fixes #18
If RabbitMQ queue has a lot of pending payloads being processed by a backend at the time of the check, the queue will appear as loaded while it’s not. We now allow to check again later after a defined ‘backend mean queue processing time’ to check if the backend is slow or if it was just an unusual payload spike. This avoids the status page from blinking from green to orange all day long for loaded queues.

Signed-off-by: Valerian Saliou <valerian@valeriansaliou.name>