Add a `enable_backoff_timer` config option to avoid spamming reminders
This is pretty simple, we increment a counter every time we send
a reminder and we multiply the `reminder_interval` by its value.

If it's disabled then the counter stays at 1 at all time so we can avoid
having a separate branch to check if we should take it into account.

Fixes #58